cambios hechos: usuario, autenticacion y reportes 07-06-2021 19-06

+++ IronHorseCore/Models/Toll.cs
@@ -10,7 +10,7 @@ namespace IronHorseCore.Models
         public int Id { get; set; }
         public int OperationsId { get; set; }
         public DateTime DatePay { get; set; }
-        public decimal Pay { get; set; }
+        public float Pay { get; set; }
 
         public virtual Operation Operations { get; set; }
     }
